removed infoPart,container and the particles from all pages, moved up to app.js if a user is logged in.
