+++ client/src/pages/detailedAvailability/detailedAvailability.js
@@ -3,13 +3,10 @@ import { Row, Col } from 'react-grid-system';
 import "./style.css";
 import api from '../../utils/api';
 import DeatiledSubmit from "../../components/detailedSubmit";
-import InfoPart from "../../components/infoPart";
 import Header from "../../components/Header";
 import moment from "moment";
-import { Container, Table } from 'react-bootstrap';
-import Particles from "react-particles-js";
+import Table from 'react-bootstrap/Table';
 
-const particleOpt = { particles: { number: { value: 120, density: { enable: true, value_area: 1000 } } } };
 const today = moment().format("YYYY-MM-DD");
 class DetailedAvailability extends Component {
     state = {
@@ -46,58 +43,49 @@ class DetailedAvailability extends Component {
 
     render() {
         return (
-            <Container>
-                <Particles params={particleOpt} id="particul" />
-
-                <Row >
-                <Col xs={6} sm={4} md={3} lg={3} xl={2}>
-                        <InfoPart />
-                    </Col>
-                    <Col xs={6} sm={8}md={9} lg={9} xl={10}>
-                        <Row>
-                            <Col xl={12}>
-                                <Header>DETAILED AVAILABILITY</Header>
-                            </Col>
-                        </Row>
-                        <Row>
-                            <Col xl={12}>
-                                <div id="res">
-                                    <Row>
-                                        <DeatiledSubmit availableChecked={this.state.availableChecked} occupiedChecked={this.state.occupiedChecked} handleCheckbox={this.handleCheckbox} handleStartDate={this.handleStartDate} />
-                                    </Row>
-                                </div>
-                                <div id="res">
-                                    <Row style={{ paddingBottom: "20px" }}>
-                                        <Col xl={12}>
-                                            <Table>
-                                                <tbody>
-                                                    <tr>
-                                                        <th className="th" id="date">Date</th>
-                                                        {this.state.roomTypes.map(type => (
-                                                            <th key={type.room_type_id} className="th">{type.type}</th>
-                                                        ))}
-                                                        <th className="th" id="total" >Total Rooms</th>
-                                                    </tr>
-                                                    {this.state.availableRooms.map(tot => (
-                                                        <tr key={tot.date}>
-                                                            <td className="tableTD">{tot.date} ({moment(tot.date).format("dddd")})</td>
-                                                            <td className="tableTD">{this.state.availableChecked===true ? tot.AvailableType1 : tot.OccupiedType1}</td>
-                                                            <td className="tableTD">{this.state.availableChecked===true ? tot.AvailableType2 : tot.OccupiedType2}</td>
-                                                            <td className="tableTD">{this.state.availableChecked===true ? tot.AvailableType3 : tot.OccupiedType3}</td>
-                                                            <td className="tableTD">{this.state.availableChecked===true ? tot.TotalAvailable : tot.TotalOccupied}</td>
-                                                        </tr>
-                                                    ))}
-                                                </tbody>
-                                            </Table>
-                                        </Col>
-                                    </Row>
-                                </div>
-                            </Col >
-                        </Row >
-                    </Col>
-                </Row>
-            </Container >
-        )
+        <div>
+        <Row>
+            <Col xl={12}>
+                <Header>DETAILED AVAILABILITY</Header>
+            </Col>
+        </Row>
+        <Row>
+            <Col xl={12}>
+                <div id="res">
+                    <Row>
+                        <DeatiledSubmit availableChecked={this.state.availableChecked} occupiedChecked={this.state.occupiedChecked} handleCheckbox={this.handleCheckbox} handleStartDate={this.handleStartDate} />
+                    </Row>
+                </div>
+                <div id="res">
+                    <Row style={{ paddingBottom: "20px" }}>
+                        <Col xl={12}>
+                            <Table>
+                                <tbody>
+                                    <tr>
+                                        <th className="th" id="date">Date</th>
+                                        {this.state.roomTypes.map(type => (
+                                            <th key={type.room_type_id} className="th">{type.type}</th>
+                                        ))}
+                                        <th className="th" id="total" >Total Rooms</th>
+                                    </tr>
+                                    {this.state.availableRooms.map(tot => (
+                                        <tr key={tot.date}>
+                                            <td className="tableTD">{tot.date} ({moment(tot.date).format("dddd")})</td>
+                                            <td className="tableTD">{this.state.availableChecked===true ? tot.AvailableType1 : tot.OccupiedType1}</td>
+                                            <td className="tableTD">{this.state.availableChecked===true ? tot.AvailableType2 : tot.OccupiedType2}</td>
+                                            <td className="tableTD">{this.state.availableChecked===true ? tot.AvailableType3 : tot.OccupiedType3}</td>
+                                            <td className="tableTD">{this.state.availableChecked===true ? tot.TotalAvailable : tot.TotalOccupied}</td>
+                                        </tr>
+                                    ))}
+                                </tbody>
+                            </Table>
+                        </Col>
+                    </Row>
+                </div>
+            </Col >
+        </Row >
+    </div>
+    )
     }
 }
