+++ client/src/pages/reservation/reservation.js
@@ -2,11 +2,8 @@ import React, { Component } from "react";
 import api from '../../utils/api';
 import { Row, Col } from 'react-grid-system';
 import "./style.css";
-import InfoPart from "../../components/infoPart";
 import Header from "../../components/Header";
-import Particles from "react-particles-js";
 
-const particleOpt = { particles: { number: { value: 120, density: { enable: true, value_area: 1000 } } } };
 class Reservation extends Component {
     state = {
         ReservationInfo: {},
@@ -27,67 +24,56 @@ class Reservation extends Component {
     }
 
     render() {
-
         return (
-            <Particles params={particleOpt} id="particul" />
-
-            <Row id="dashboardTable1">
-                <Col xs={6} sm={4} md={3} lg={3} xl={2}>
-
-                    <InfoPart />
-                </Col>
-                <Col xs={6} sm={8} md={9} lg={9} xl={10}>                    <row>
-                    <div id="header">
-                        <Header>RESERVATION CONFIRMATION</Header>
-                    </div>
-                    <div id="res">
-                        <table>
-                            <tr>
-                                <th>Reservation ID</th>
-                                <th>Created On</th>
-                                <th>Name</th>
-                                <th>Email</th>
-                                <th>Phone</th>
-                                <th>CC Last 4</th>
-                                <th>User ID</th>
+            <div>
+                <div id="header">
+                    <Header>RESERVATION CONFIRMATION</Header>
+                </div>
+                <div id="res">
+                    <table>
+                        <tr>
+                            <th>Reservation ID</th>
+                            <th>Created On</th>
+                            <th>Name</th>
+                            <th>Email</th>
+                            <th>Phone</th>
+                            <th>CC Last 4</th>
+                            <th>User ID</th>
+                        </tr>
+                        <tr>
+                            <td>{this.state.ReservationInfo.reservation_id}</td>
+                            <td>{this.state.ReservationInfo.created_at}</td>
+                            <td>{this.state.ReservationInfo.last_name}, {this.state.ReservationInfo.first_name}</td>
+                            <td>{this.state.ReservationInfo.email}</td>
+                            <td>{this.state.ReservationInfo.phone}</td>
+                            <td>{this.state.ReservationInfo.ccLastFour}</td>
+                            <td>{this.state.ReservationInfo.user_id}</td>
+                        </tr>
+                    </table>
+                    <table>
+                        <tr>
+                            <th>Confirmation Number</th>
+                            <th>Arrival Date</th>
+                            <th>Departure Date</th>
+                            <th>Room Type</th>
+                            <th>Room Rate</th>
+                            <th>Adults</th>
+                            <th>Comments</th>
+                        </tr>
+                        {this.state.RoomInfo.map(room => (
+                            <tr key={room.res_room_id}>
+                                <td>{room.confirmation_code}</td>
+                                <td>{room.check_in_date}</td>
+                                <td>{room.check_out_date}</td>
+                                <td>{room.type}</td>
+                                <td>{room.rate}</td>
+                                <td>{room.adults}</td>
+                                <td>{room.comments}</td>
                             </tr>
-                            <tr>
-                                <td>{this.state.ReservationInfo.reservation_id}</td>
-                                <td>{this.state.ReservationInfo.created_at}</td>
-                                <td>{this.state.ReservationInfo.last_name}, {this.state.ReservationInfo.first_name}</td>
-                                <td>{this.state.ReservationInfo.email}</td>
-                                <td>{this.state.ReservationInfo.phone}</td>
-                                <td>{this.state.ReservationInfo.ccLastFour}</td>
-                                <td>{this.state.ReservationInfo.user_id}</td>
-                            </tr>
-                        </table>
-
-                        <table>
-                            <tr>
-                                <th>Confirmation Number</th>
-                                <th>Arrival Date</th>
-                                <th>Departure Date</th>
-                                <th>Room Type</th>
-                                <th>Room Rate</th>
-                                <th>Adults</th>
-                                <th>Comments</th>
-                            </tr>
-                            {this.state.RoomInfo.map(room => (
-                                <tr key={room.res_room_id}>
-                                    <td>{room.confirmation_code}</td>
-                                    <td>{room.check_in_date}</td>
-                                    <td>{room.check_out_date}</td>
-                                    <td>{room.type}</td>
-                                    <td>{room.rate}</td>
-                                    <td>{room.adults}</td>
-                                    <td>{room.comments}</td>
-                                </tr>
-                            ))}
-                        </table>
-                    </div>
-                </row>
-                </Col >
-            </Row >
+                        ))}
+                    </table>
+                </div>
+            </div>
         )
     }
 }
